Addedfiles and changes required for deployment to heroku

+++ bot/models.py
@@ -5,9 +5,8 @@ from django.utils import timezone
 class contact_us(models.Model):
     name = models.CharField(max_length=70)
     email=models.EmailField(default="aaa@bb.c")
-    phone=models.IntegerField(default=10)
+    phone=models.BigIntegerField(default=10)
     message=models.CharField(max_length=500)
-
     date=models.DateField(auto_now=True)
     time=models.TimeField(auto_now=True)
